requirements.txt
fastapi
uvicorn[standard]
kafka-python
requests
tos
pydantic
tenacity
python-dotenv
